Move mock_config_entry fixture to conftest for cross-platform reuse
